mod transfer param from int64 to asset (run failed)

+++ libraries/chain/wasm_price.cpp
@@ -517,7 +517,7 @@ namespace contento { namespace chain { namespace wasm_price {
     // contract_balance_api
     WASM_PRICE_FN_CONSTS(
                          (get_contract_balance,      (int64_t) )
-                         (transfer,                  (void*, int64_t, int64_t))
+                         (transfer,                  (void*, int64_t, int))
                          (get_value,                 (int64_t))
                          (pay_prohibited,            (void*))
                          );
